# Optimize RapidJSON benchmark