@stephenberry 2022-12-05
Thanks for this pull request! I optimized it more by eliminating a string allocation.